backend/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.0
opencv-python==4.12.0.88
Pillow==11.0.0
pytesseract==0.3.10
mss==9.0.1
numpy==2.1.2
python-socketio==5.9.0
pyautogui==0.9.54
pynput==1.7.6
requests==2.32.3